fix(test): move nonDeliverableTerminalTurn tests to unique filename to avoid merge conflict
agent-runner-execution-terminal-failures.test.ts exists in main (722 lines) and
was also new in PR, causing a both-sides-added conflict that GitHub cannot
auto-resolve. Move the 3 new tests to a uniquely-named file
(agent-runner-execution-nondeliverable.test.ts) that doesn't exist in main,
eliminating the conflict entirely.
